Rename master term model source files.
