+++ src/gui/models/commonImproperModel.h
@@ -3,12 +3,13 @@
 
 #pragma once
 
-#include "classes/speciesBond.h"
-#include "gui/models/masterTermModel.h"
+#include "classes/speciesImproper.h"
+#include "gui/models/commonTermModel.h"
 #include "gui/models/modelUpdater.h"
+#include "templates/optionalRef.h"
 
-// CommonBond model
-class CommonBondModel : public CommonTermModel
+// CommonImproper model
+class CommonImproperModel : public CommonTermModel
 {
     Q_OBJECT
 
@@ -19,11 +20,11 @@ class CommonBondModel : public CommonTermModel
     void modelsUpdated();
 
     public:
-    explicit CommonBondModel(Species *species);
+    explicit CommonImproperModel(Species *species);
 
     private:
     // Source term data
-    std::vector<std::shared_ptr<CommonBond>> &sourceData_;
+    std::vector<std::shared_ptr<CommonImproper>> &sourceData_;
 
     public:
     // Refresh model data
@@ -36,9 +37,7 @@ class CommonBondModel : public CommonTermModel
     int rowCount(const QModelIndex &parent = QModelIndex()) const override;
     QVariant getTermData(int row, CommonTermModelData::DataType dataType) const override;
     bool setTermData(int row, CommonTermModelData::DataType dataType, const QVariant &value) override;
-    const std::shared_ptr<CommonBond> &rawData(const QModelIndex &index) const;
-    bool insertRows(int row, int count, const QModelIndex &parent) override;
-    bool removeRows(int row, int count, const QModelIndex &parent) override;
+    const std::shared_ptr<CommonImproper> &rawData(const QModelIndex &index) const;
 
     private:
     ModelUpdater modelUpdater;
